Minor Corrections
Changed sort order + corrected a parameter call.

App is more or less complete.
The history is not loading correctly when switching pages or first loading the app. I hope to fix this before midnight's submission.

I was also hoping to add a delete/clear option for the player history but due to time constrains I may not complete this feature.